+++ src/modules/gfx/drawable.c
@@ -96,7 +96,7 @@ void vbuffer_free(void *buf) {
  *    @return void *    The mesh.
  */
 void *mesh_create(void *v) {
-    size_t i;
+    unsigned int i;
     mesh_t       *mesh = (mesh_t *)malloc(sizeof(mesh_t));
 
     if (mesh == (mesh_t *)0x0) {
@@ -139,12 +139,12 @@ void mesh_set_vbuffer(void *m, void *v) {
  *
  *    @param void *m              The mesh.
  *    @param void *a              The asset.
- *    @param size_t size   The size of the asset.
+ *    @param unsigned int size   The size of the asset.
  */
-void mesh_append_asset(void *m, void *a, size_t size) {
-    size_t j;
-    size_t index;
-    size_t offset = 8 * CHIK_GFX_DRAWABLE_MESH_MAX_ASSETS;
+void mesh_append_asset(void *m, void *a, unsigned long size) {
+    unsigned int j;
+    unsigned int index;
+    unsigned int offset = 8 * CHIK_GFX_DRAWABLE_MESH_MAX_ASSETS;
 
     if (m == (void *)0x0) {
         LOGF_ERR("Mesh is null.\n");
@@ -167,8 +167,8 @@ void mesh_append_asset(void *m, void *a, size_t size) {
 
     auto value = _heapchk();
 
-    memcpy((void *)((size_t)mesh->assets + mesh->assets_count * 8), &mesh->assets_size, sizeof(mesh->assets_size));
-    memcpy((void *)((size_t)mesh->assets + mesh->assets_size), a, size);
+    memcpy((void *)((unsigned int)mesh->assets + mesh->assets_count * 8), &mesh->assets_size, sizeof(mesh->assets_size));
+    memcpy((void *)((unsigned int)mesh->assets + mesh->assets_size), a, size);
     mesh->assets_size += size;
     mesh->assets_count++;
 }
@@ -178,12 +178,12 @@ void mesh_append_asset(void *m, void *a, size_t size) {
  *
  *    @param void *m              The mesh.
  *    @param void *a              The asset.
- *    @param size_t size   The size of the asset.
- *    @param size_t i      The index of the asset.
+ *    @param unsigned int size   The size of the asset.
+ *    @param unsigned int i      The index of the asset.
  */
-void mesh_set_asset(void *m, void *a, size_t size, size_t i) {
-    size_t j;
-    size_t offset;
+void mesh_set_asset(void *m, void *a, unsigned long size, unsigned long i) {
+    unsigned int j;
+    unsigned int offset;
 
     if (m == (void *)0x0) {
         LOGF_ERR("Mesh is null.\n");
@@ -196,28 +196,28 @@ void mesh_set_asset(void *m, void *a, size_t size, size_t i) {
 
     mesh_t *mesh = (mesh_t *)m;
 
-    offset = *(size_t *)((size_t)mesh->assets + i * 8);
+    offset = *(unsigned int *)((unsigned int)mesh->assets + i * 8);
 
-    memcpy((void *)((size_t)mesh->assets + offset), a, size);
+    memcpy((void *)((unsigned int)mesh->assets + offset), a, size);
 }
 
 /*
  *    Returns the data of an asset.
  *
  *    @param void *a            The assets.
- *    @param size_t i    The index of the asset.
+ *    @param unsigned int i    The index of the asset.
  *
  *    @return void *            The asset data.
  */
-void *mesh_get_asset(void *a, size_t i) {
+void *mesh_get_asset(void *a, unsigned long i) {
     if (a == (void *)0x0) {
         LOGF_ERR("Assets are null.\n");
         return (void *)0x0;
     }
 
-    size_t offset = *(size_t *)((size_t)a + i * 8);
+    unsigned int offset = *(unsigned int *)((unsigned int)a + i * 8);
 
-    return (void *)((size_t)a + offset);
+    return (void *)((unsigned int)a + offset);
 }
 
 /*
